moved filed to component level
