+++ src/components/avatar/index.test.tsx
