+++ src/components/avatar/styles.tsx
