+++ src/components/avatar/types.ts
