+++ src/components/latest_blocks_desktop/index.test.tsx
@@ -21,8 +21,8 @@ describe('LatestBlocksDesktop', () => {
           }
         }
         data={[{
-          proposer: () => <div>proposer</div>,
-          height: () => <div>1,234</div>,
+          proposer: <div>proposer</div>,
+          height: <div>1,234</div>,
           time: '10 Jan 2020,13:00:22 UTC',
           hash: '89832B67F594asddw32',
           tx: 2,
