+++ src/components/latest_blocks_desktop/styles.tsx
