+++ src/components/latest_blocks_mobile/styles.tsx
