+++ src/components/latest_blocks_mobile/types.ts
