+++ src/components/latest_blocks_mobile/utils.tsx
