+++ src/components/stabilities/styles.tsx
