+++ src/components/table_preview_wrapper/styles.tsx
