+++ stories/utils/data_block.tsx
@@ -1,4 +1,4 @@
-import React, { PureComponent } from 'react';
+import React from 'react';
 import {
   AreaChart,
   Area,
